Implement include_str! for tags
